Move LoginViewModel into ui.sync package
Relocate LoginViewModel.kt from ui/sync/viewmodel/ to ui/sync/ alongside
LoginActivity. Update its package declaration to
org.ole.planet.myplanet.ui.sync and drop the now-redundant import in
LoginActivity (same package).
